// src/app/models/people.ts
export interface Person {
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  homeworld: string;
  films: string[];
  species: string[];
  vehicles: string[];
  starships: string[];
  created: string;
  edited: string;
  url: string;
}

export interface PeopleWitBMI extends Person {
  BMI: number;
}

export interface GroupPeopleBy {
  fromAge: number;
  toAge: number;
}